+++ modules/apps/foundation/frontend-taglib/frontend-taglib-clay/.babelrc
@@ -1,3 +1,3 @@
 {
-	"presets": ["es2015", "liferay-project"]
+	"presets": ["env", "liferay-project"]
 }
