+++ modules/apps/foundation/frontend-taglib/frontend-taglib-clay/package.json
@@ -26,15 +26,19 @@
 	},
 	"devDependencies": {
 		"babel-cli": "6.26.0",
-		"babel-preset-es2015": "6.24.1",
+		"babel-preset-env": "^1.6.1",
 		"babel-preset-liferay-project": "1.2.2",
+		"check-source-formatting": "^2.0.4",
 		"liferay-npm-bundler": "1.2.2",
 		"liferay-npm-bundler-preset-standard": "^1.2.2"
 	},
 	"name": "frontend-taglib-clay",
 	"private": true,
 	"scripts": {
-		"build": "babel --source-maps -d classes/META-INF/resources src/main/resources/META-INF/resources && liferay-npm-bundler"
+		"build": "babel --source-maps -d classes/META-INF/resources src/main/resources/META-INF/resources && liferay-npm-bundler",
+		"checkFormat": "npm run csf",
+		"csf": "csf src/**/*.es.js",
+		"format": "npm run csf -- -i"
 	},
 	"version": "1.0.0"
 }
